requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
databases
psycopg2-binary
pydantic
python-dotenv
